Rewrite for better parallelization, miscellaneous improvements to plotting options (#11)
The main change of this PR is a re-arranging of the logic flow to improve parallelization. We now have a dedicated setupargs() function called before plotit(): now instead of looping over variables and levels within plotit(), all sets of arguments for each user-specified plot are generated in advance, and passed to Python multiprocessing for efficient distribution of tasks.

Other improvements include:

A new color option for coastline/political boundaries
Only create the selected Cartopy projection rather than all; this doesn't save much in the way of resources, but it does fix some edge case errors such as with polar plotting