New major overhaul. Currently in separate script (custom_plot.py), but current thinking is this will supersede old plotting script.
This script maintains the same core functionality, but the major difference will be that each
variable to be plotted will have its own dictionary of settings. This will enable the specification
of different plotting settings for each variable, but most importantly, will also allow for the
definition of custom variables; e.g. calculating and plotting the 6-hour precipitation from a total
accumulation field. The logic of file reads has also been completely reorganized to only retain the
variables that we need in a single data structure (to cut down on memory usage) and also greatly
speed up plotting across multiple input files.

These features are only partially implemented in this initial version, but will be fully
implemented over the next several commits. The current version is a proof of concept that only
plots hard-coded variables from the input files; the next step will be fully implementing custom
variables.