Bugfix for input data file names (#13)
Currently, if a list of data file names is specified in the config file, the script will crash.  This fixes that issue and also allows either a glob-able string or a list of glob-able strings to be specified as the (set of) data files in the config file to search and plot.