New derived variable calculations, new custom colormaps, better memory performance, other improvements (#15)
This is a combination of many rounds of improvements, including changes from #14 which has been superseded. Several features have been added based on user requests, and other quality-of-life improvements that I have included for my own purposes over the last few months. The only functionality not carried forward from the previous version of the script (but can be added in the future if needed) is the ability to plot "all" variables; now users must explicitly list all variables to plot ("all levels" plotting is still available).

**NOTE: the old config.yaml format is deprecated, so old config files will need to be updated for this new format.** Many examples of this new format have been added to the GitHub wiki: https://github.com/dtcenter/mpas_plot/wiki/mpas_plot-config-examples-and-use-cases